Use BaseTaskFactory
