+++ src/aiida_workgraph/tasks/factory/aiida_task.py
@@ -1,13 +1,13 @@
 from __future__ import annotations
-from copy import deepcopy
 from aiida_workgraph.task import Task
-from node_graph.utils import list_to_dict
 from typing import Any, Callable, Dict, List, Optional, Tuple, Union
 from aiida_workgraph.orm.mapping import type_mapping
 from node_graph.executor import NodeExecutor
 import inspect
 from aiida.engine.processes.ports import PortNamespace
 from aiida_workgraph.config import builtin_inputs, builtin_outputs
+from .base import BaseTaskFactory
+from aiida_workgraph.utils import validate_task_inout
 
 
 def add_input_recursive(
@@ -158,11 +158,8 @@ def get_task_data_from_aiida_component(
     return tdata
 
 
-class AiiDAComponentTaskFactory:
-    """
-    A factory to create specialized subclasses of Task,
-    embedding the 'ndata' (i.e., all relevant data).
-    """
+class AiiDAComponentTaskFactory(BaseTaskFactory):
+    """Factory for creating tasks from AiiDA components."""
 
     @classmethod
     def from_aiida_component(
@@ -182,6 +179,8 @@ class AiiDAComponentTaskFactory:
 
         identifier = identifier or callable.__name__
         error_handlers = error_handlers or []
+        inputs = validate_task_inout(inputs or [], "inputs")
+        outputs = validate_task_inout(outputs or [], "outputs")
 
         tdata = get_task_data_from_aiida_component(
             callable, inputs=inputs, outputs=outputs
@@ -193,64 +192,3 @@ class AiiDAComponentTaskFactory:
         TaskCls = cls(tdata)
         callable.identifier = identifier
         return TaskCls
-
-    def __new__(cls, ndata: dict):
-        """
-        Instead of returning an instance of AiiDAComponentTaskFactory,
-        we return the new Task subclass that uses `ndata`.
-        """
-
-        class _AiiDAComponentTask(Task):
-            """A specialized Task with the embedded ndata."""
-
-            _ndata = deepcopy(ndata)
-
-            def __init__(self, *args, **kwargs):
-                self.identifier = self._ndata["identifier"]
-                self.node_type = self._ndata.get("metadata", {}).get(
-                    "node_type", "NORMAL"
-                )
-                self.catalog = self._ndata.get("metadata", {}).get("catalog", "Others")
-                self._executor = NodeExecutor(**self._ndata["executor"])
-                self._error_handlers = self._ndata.get("error_handlers", [])
-                super().__init__(*args, **kwargs)
-
-            def create_properties(self):
-                properties = list_to_dict(self._ndata.get("properties", {}))
-                for prop in properties.values():
-                    self.add_property(
-                        prop.pop("identifier", type_mapping["default"]), **prop
-                    )
-
-            def create_sockets(self):
-                inputs = list_to_dict(self._ndata.get("inputs", {}))
-                for inp in inputs.values():
-                    if isinstance(inp, str):
-                        inp = {"identifier": type_mapping["default"], "name": inp}
-                    kwargs = {}
-                    if "property_data" in inp:
-                        kwargs["property_data"] = inp.pop("property_data")
-                    if "sockets" in inp:
-                        kwargs["sockets"] = inp.pop("sockets")
-                    self.add_input(
-                        inp.get("identifier", type_mapping["default"]),
-                        name=inp["name"],
-                        metadata=inp.get("metadata", {}),
-                        link_limit=inp.get("link_limit", 1),
-                        **kwargs,
-                    )
-
-                outputs = list_to_dict(self._ndata.get("outputs", {}))
-                for out in outputs.values():
-                    if isinstance(out, str):
-                        out = {"identifier": type_mapping["default"], "name": out}
-                    self.add_output(
-                        out.get("identifier", type_mapping["default"]),
-                        name=out["name"],
-                        metadata=out.get("metadata", {}),
-                    )
-
-            def get_executor(self):
-                return self._ndata.get("executor", None)
-
-        return _AiiDAComponentTask
